feature: receber dados da tela de configuração de horários, criar lógica para exibir as doses separadamentes com seus respectivos horários selecionados e importar os estilos da página
