+++ meu-app/frontend/app/assets/styles/addInsulin.js
@@ -1,0 +1,41 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    paddingTop: 16,
+    paddingHorizontal: 20,
+    gap: 8,
+    backgroundColor: "#FDFDFD",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  scrollContainer: {
+    flexGrow: 1,
+    gap: 16,
+  },
+  notificacaoContainer: {
+    width: 320,
+    height: 97,
+    gap: 16,
+  },
+  notificacaoTitle: {
+    fontFamily: "Lato_400Regular",
+    fontSize: 14,
+    lineHeight: 21,
+  },
+  textNotificacao: {
+    fontFamily: "Lato_400Regular",
+    fontSize: 16,
+    lineHeight: 22,
+  },
+  frequencia: {
+    color: "#2F39D3",
+  },
+  dados: {
+    fontFamily: "Lato_400Regular",
+    fontSize: 14,
+    color: "#282828",
+  },
+});
+
+export default styles;
